Create 1VariablesAndValues
Notes are written in R with natural language prefaced with the # marker. Though you can run these files and its contents into an R terminal, this specific files' code lacks any real values and data for the variables and value functions. I abstracted as much as I could early on, though in the later chapters of the lesson, I do not abstract and instead use copious amounts of notes to breakdown and define what is happening in the code. 

Again, these are all notes and not actual code, though code is contained within the notes.